flask_app/requirements.txt
Flask
celery[redis]
mysql-connector